cart_service/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.6
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-redis==5.4.0
djangorestframework==3.14.0
drf-spectacular==0.26.5
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.30.2
rpds-py==0.10.6
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
